Simplify SSM POST endpoint
This removes the need to pass namespace_id, since it can be inferred
from the project in the URL path. It also changes from an array of
arbirary params to an object of environment -> value pairs, since
a project will only have one value per name per environment. It
also changes the meaning of the name field, instead of being the
entire path, it's the suffix after the pre-defined project prefix.
The prefix is generated on the backend now depending on the project.
This ensures the endpoint is project-oriented and not a generic
SSM POST endpoint.