Update SSM GET response schema
This updates the array of params to an object with environment keys
to ensure one param per environment. It also adds the self key which
is a URL that can be used to refer to a param with a specific name,
which will be used in the upcoming PATCH endpoint.