Add Operations Log occurred_at
This field tracks *when* the event occured at as opposed to when it was
added to the system. The recorded at field tracks when it was added.

I also removed the ability to set recorded_at and recorded_by since they
are set automatically by the API. We specifically DO NOT want the user
to be able to set these.